split out read json file from github stuff
